agentic_ai_projects/simple_chatboot/pyproject.toml
[project]
name = "simple-chatboot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.9",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
]
